Reorganize ts/components directory
