+++ optuna_dashboard/ts/components/Artifact/ArtifactCardMedia.tsx
