+++ optuna_dashboard/ts/components/Artifact/ThreejsArtifactViewer.tsx
