+++ optuna_dashboard/ts/components/Artifact/TrialArtifactCards.tsx
@@ -20,7 +20,7 @@ import DownloadIcon from "@mui/icons-material/Download"
 import DeleteIcon from "@mui/icons-material/Delete"
 import FullscreenIcon from "@mui/icons-material/Fullscreen"
 
-import { actionCreator } from "../action"
+import { actionCreator } from "../../action"
 import { useDeleteTrialArtifactDialog } from "./DeleteArtifactDialog"
 import {
   useThreejsArtifactModal,
