+++ optuna_dashboard/ts/components/Preferential/PreferentialAnalytics.tsx
@@ -8,10 +8,10 @@ import {
   useTheme,
 } from "@mui/material"
 import Grid2 from "@mui/material/Unstable_Grid2"
-import { DataGrid, DataGridColumn } from "./DataGrid"
-import { BestTrialsCard } from "./BestTrialsCard"
-import { useStudyDetailValue, useStudySummaryValue } from "../state"
-import { Contour } from "./GraphContour"
+import { DataGrid, DataGridColumn } from "../DataGrid"
+import { BestTrialsCard } from "../BestTrialsCard"
+import { useStudyDetailValue, useStudySummaryValue } from "../../state"
+import { Contour } from "../GraphContour"
 
 export const PreferentialAnalytics: FC<{ studyId: number }> = ({ studyId }) => {
   const theme = useTheme()
